+++ src/main/java/com/lesniewicz/api/dto/LanguageDto.java
@@ -2,7 +2,7 @@ package com.lesniewicz.api.dto;
 
 import java.time.OffsetDateTime;
 
-public record LanguageDto(Integer languageId,
+public record LanguageDto(Long languageId,
                           String name,
                           OffsetDateTime lastUpdate) {
 }
